gcc-2.7.2-t800.5
This is a bugfix release: running out of spill registers in reload on
certain programs using asms has been cured (I hope).

Note that one asm example in README.T800 turned out to be incorrect.

  void
  outword (unsigned word, unsigned chan)
  {
    asm volatile
        ("outword"
         : /* no outputs */
         : "bP" (chan), "aP" (word)
         : "Creg", "FAreg", "FBreg", "FCreg", "Wreg[0]");
  }

should actually be written as

  void
  outword (unsigned word, unsigned chan)
  {
    asm volatile
        ("outword"
         : /* no outputs */
         : "bP" (chan), "aP" (word)
         : "Areg", "Breg", "Creg", "FAreg", "FBreg", "FCreg", "Wreg[0]");
  }

or GCC may attempt to keep values in Areg or Breg across the asm,
which is incorrect.

Thanks to Roman Pozlevich <roma@botik.ru> for discovering the problem.